+++ src/rer/solrpush/utils/solr_indexer.py
@@ -130,7 +130,7 @@ def can_index(item):
     return is_right_portal_type(item)
 
 
-def create_index_dict(item):
+def create_index_dict(item, default={}):
     """Restituisce un dizionario pronto per essere 'mandato' a SOLR per
     l'indicizzazione.
     """
@@ -148,7 +148,13 @@ def create_index_dict(item):
             # repsonses and can be copied in solr configuration.
             continue
         field_type = field_infos.get("type")
-        value = getattr(adapter, field, None)
+        try:
+            value = getattr(adapter, field, None)
+        except Exception:
+            # if we are in creation and indexing a content with a File, SearchableText
+            # raise PosKeyError because it tries to read the file too, but the blob
+            # isn't created yet.
+            value = default.get(field, None)
         if not value and value is not False:
             continue
         if callable(value):
@@ -254,7 +260,7 @@ def push_to_solr(item_or_obj):
             context = obj
     if not can_index(context):
         return False
-    index_dict = create_index_dict(context)
+    index_dict = create_index_dict(item=context, default=item_or_obj)
     attachment = None
     if index_dict.get("attachment", None):
         attachment = index_dict.pop("attachment", None)